take 2nd review : create program customizer + fixes
