+++ libs/common/src/questionnaire/types/types.ts
@@ -28,11 +28,11 @@ export enum StructureSize {
   ETI_GE = 'ETI_GE'
 }
 
-export const StructureSizeMap: { [key in StructureSize]: number } = {
+export const SizeToWorkforce: { [key in StructureSize]: number } = {
   [StructureSize.EI]: 1,
-  [StructureSize.TPE]: 10,
-  [StructureSize.PE]: 35,
-  [StructureSize.ME]: 125,
+  [StructureSize.TPE]: 19,
+  [StructureSize.PE]: 49,
+  [StructureSize.ME]: 249,
   [StructureSize.ETI_GE]: 251
 }
